Comment out celery from the requirements
The version of celery being used is really old. When attempting to 
install, the following error occurs:

    error in anyjson setup command: use_2to3 is invalid

The issue is that setuptools dropped use_2to3 at some point, so now 
anyjson (which hasn't been updated since 2012) doesn't build. I'm sure 
the most recent versions of Celery don't rely on it, but upgrading from 
Celery 3 to 4 is non-trivial (and I'm not even sure Shareabouts need 
it).